+++ packages/fleury_widgets/lib/src/message_list.dart
@@ -149,7 +149,7 @@ class MessageListController extends Notifier {
   }
 
   bool get isFollowing => _list.isFollowing;
-  bool get atBottom => _list.atBottom;
+  bool get atBottom => _list.atEnd;
   int get unseenCount => _list.unseenCount;
 
   ({int first, int last})? get visibleRange => _list.visibleRange;
@@ -164,7 +164,7 @@ class MessageListController extends Notifier {
   void scrollToBottom() {
     _checkNotDisposed();
     _list.followTail = true;
-    _list.jumpToBottom();
+    _list.jumpToEnd();
   }
 
   void _checkNotDisposed() {